requirements.txt
torch>=1.13.0
transformers>=4.30.0
datasets>=2.14.0
vllm>=0.1.0
numpy>=1.21.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.64.0
pandas>=1.3.0
